final finish

+++ setupgit.sh
@@ -9,7 +9,7 @@ cd nasa-judge-api
 
 for a in $(ls -a)
 do
-    if [ "$a" != "." ] && [ "$a" != ".." ] && [ "$a" != ".git" ] && [ "$a" != "README.md" ] && [ "$a" != "install.sh" ] && [ "$a" != "remove.sh" ] && [ "$a" != "setupdefaultiptables.sh" ] && [ "$a" != "iptablesconf.conf" ] && [ "$a" != "setupgit.sh" ] && [ "$a" != "db.conf" ] && [ "$a" != "exfrontenddata.json" ]
+    if [ "$a" != "." ] && [ "$a" != ".." ] && [ "$a" != ".git" ] && [ "$a" != "README.md" ] && [ "$a" != "install.sh" ] && [ "$a" != "remove.sh" ] && [ "$a" != "setupdefaultiptables.sh" ] && [ "$a" != "iptablesconf.conf" ] && [ "$a" != "setupgit.sh" ] && [ "$a" != "db.conf" ] && [ "$a" != "finaluserlist.conf" ] && [ "$a" != "exfrontenddata.json" ]
     then
         rm -rf $a
     fi
@@ -17,7 +17,7 @@ done
 
 for a in $(ls -a /etc/nasajudgeapi)
 do
-    if [ "$a" != "." ] && [ "$a" != ".." ] && [ "$a" != "db.conf" ] && [ "$(cat /etc/nasajudgeapi/.gitignore | sed 's/\/.*//g' | sed '/^!.*/d' | grep -P "^$(echo "$a" | sed 's/\./\\\./g')$")" == "" ]
+    if [ "$a" != "." ] && [ "$a" != ".." ] && [ "$a" != "db.conf" ] && [ "$a" != "finaluserlist.conf" ] && [ "$(cat /etc/nasajudgeapi/.gitignore | sed 's/\/.*//g' | sed '/^!.*/d' | grep -P "^$(echo "$a" | sed 's/\./\\\./g')$")" == "" ]
     then
         sudo cp -r /etc/nasajudgeapi/$a $a
     fi
